feat: add containsBigInt function
